Destroy script

+++ scripts/support/destroy-system.sh
@@ -1,15 +1,36 @@
 # --------------------------------------------------------------
-# Destroys the TEST system
+# Destroys the given system
+# Takes two parameters:
+# $1 - deployment mode: 'live', 'test' or 'dev'
+# $2 - local mode (will perform cleanup): true or not set
+# Optionally, with third parameter set to 'true', will force destruction
 # Be careful: this script may destroy infrastructure seen by customers!
 # --------------------------------------------------------------
+
+if [[ $1 != "test" ]] && [[ $1 != "dev" ]]
+then
+  echo "Invalid deployment mode $1"
+  exit 1
+fi
+
+# Get system URL, e.g. test.flox.polygon-project.ch
+url="$1.$project.polygon-project.ch"
+
+# Check whether selected deployment is online (otherwise fail if 'force' is not set to true)
+online_status=$(curl -s --head "https://$url" | grep '200')
+if ! [[ $online_status || $3 == "true" ]]
+then
+  echo "Deployment in mode $1 is not online! Use 'force' to force destruction anyways."
+  exit 1
+fi
 
 # ==========================================
 # ===  Step 0: Pre-setup (Cognito, DNS)  ===
 # ==========================================
 
 # Create flox.tfvars file from flox.config.json in frontend & backend
-zsh create-flox-tfvars.sh "test"
-echo "type=\"test\"" >> flox.tfvars
+zsh create-flox-tfvars.sh "$1"
+echo "type=\"$1\"" >> flox.tfvars
 
 cd ../aws-initial-setup/0_pre-setup || exit
 
@@ -17,7 +38,7 @@ cd ../aws-initial-setup/0_pre-setup || exit
 project=$(jq '.general.project' ../../../backend/flox.config.json)
 project=${project:1:-1}
 
-frontend_build_mode=$(jq ".general.mode_test" ../../../frontend/flox.config.json)
+frontend_build_mode=$(jq ".general.mode_$1" ../../../frontend/flox.config.json)
 frontend_build_mode=${frontend_build_mode:1:-1}
 
 aws_region=$(jq '.general.aws_region' ../../../backend/flox.config.json)
@@ -29,24 +50,14 @@ organisation=${organisation:1:-1}
 # Serverless mode (API only)
 serverless_api=$(jq ".infrastructure_$1.serverless_api" ../../../backend/flox.config.json)
 
-# Replace 'TYPE' in config.tf with actual type (live, test)
-sed -i -e "s/##TYPE##/test/g" config.tf
+# Replace 'TYPE' in config.tf with actual type (dev, test)
+sed -i -e "s/##TYPE##/$1/g" config.tf
 
 # Replace 'PROJECT' in config.tf with actual project name
 sed -i -e "s/##PROJECT##/$project/g" config.tf
 
 # Replace 'ORGANISATION' in config.tf with actual organisation name
 sed -i -e "s/##ORGANISATION##/$organisation/g" config.tf
-
-# Get mode-dependent base URL
-if [[ $1 == "live" ]]
-then
-  url=$(jq ".general.live_domain" ../../../backend/flox.config.json)
-  url=${url:1:-1}
-else
-  # E.g. test.flox.polygon-project.ch
-  url="$1.$project.polygon-project.ch"
-fi
 
 # Add domain config to flox.tfvars
 echo "# ======== Domain Config ========" >> ../../support/flox.tfvars
@@ -81,8 +92,7 @@ echo "user_pool_id=\"$user_pool_id\"" >> ../../support/flox.tfvars
 echo "user_pool_client_id=\"$user_pool_client_id\"" >> ../../support/flox.tfvars
 
 # ==========================================
-# =====   Step 1: Parent DNS setup    ======
-# =====  (Applies only in TEST mode)  ======
+# =====   Step 1: Parent DNS state    ======
 # ==========================================
 cd ../1_parent-setup || exit
 
@@ -99,13 +109,13 @@ sed -i -e "s/##ORGANISATION##/$organisation/g" config.tf
 terraform refresh -var-file="../../support/flox.tfvars"
 
 # ==========================================
-# ======     Step 2: Main setup     ========
+# ======     Step 2: Main state     ========
 # ==========================================
 
 cd ../2_main-setup || exit
 
-# Replace 'TYPE' in config.tf with actual type (live, test)
-sed -i -e "s/##TYPE##/test/g" config.tf
+# Replace 'TYPE' in config.tf with actual type (dev, test)
+sed -i -e "s/##TYPE##/$1/g" config.tf
 
 # Replace 'PROJECT' in config.tf with actual project name
 sed -i -e "s/##PROJECT##/$project/g" config.tf
@@ -143,13 +153,16 @@ terraform destroy -auto-approve -var-file="../../support/flox.tfvars"
 
 # ==========================================
 # ======      Step 4: Cleanup       ========
+# ======    (only in local mode)    ========
 # ==========================================
+if [[ $2 == 'true' ]]
+then
+  # Remove .zip files
+  rm -f ../2_main-setup/frontend.zip
+  rm -f ../2_main-setup/backend.zip
 
-# Remove .zip files
-rm -f ../2_main-setup/frontend.zip
-rm -f ../2_main-setup/backend.zip
-
-# Reset all config.tf files to their respective template files
-cp ../0_pre-setup/config.tftemplate ../0_pre-setup/config.tf
-cp ../1_parent-setup/config.tftemplate ../1_parent-setup/config.tf
-cp ../2_main-setup/config.tftemplate ../2_main-setup/config.tf
+  # Reset all config.tf files to their respective template files
+  cp ../0_pre-setup/config.tftemplate ../0_pre-setup/config.tf
+  cp ../1_parent-setup/config.tftemplate ../1_parent-setup/config.tf
+  cp ../2_main-setup/config.tftemplate ../2_main-setup/config.tf
+fi
